% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_rna-sga-coma.R
\name{alg_1sided}
\alias{alg_1sided}
\alias{rna}
\alias{sga}
\alias{sgaplus}
\alias{coma}
\title{Algorithms for Optimum Sample Allocation Under One-Sided Bound Constraints}
\usage{
rna(
  total_cost,
  A,
  bounds = NULL,
  unit_costs = 1,
  cmp = .Primitive(">="),
  details = FALSE
)

sga(total_cost, A, M)

sgaplus(total_cost, A, M)

coma(total_cost, A, M)
}
\arguments{
\item{total_cost}{(\code{numeric(1)})\cr total survey cost \eqn{c}.
Must be strictly positive.
Additionally:
\itemize{
\item If one-sided lower bounds \eqn{m_1, \ldots, m_H} are imposed, it is
required that \eqn{c \geq \sum_{h=1}^H c_h m_h}, i.e.
\code{total_cost >= sum(unit_costs * bounds)}.
\item If one-sided upper bounds \eqn{M_1, \ldots, M_H} are imposed, it is
required that \eqn{c \leq \sum_{h=1}^H c_h M_h}, i.e.
\code{total_cost <= sum(unit_costs * bounds)}.
}}

\item{A}{(\code{numeric})\cr population constants \eqn{A_1,\ldots,A_H}.
All values must be strictly positive.}

\item{bounds}{(\code{numeric} or \code{NULL})\cr optional lower bounds \eqn{m_1,\ldots,m_H},
or upper bounds \eqn{M_1,\ldots,M_H}, or \code{NULL} to indicate that no inequality
constraints are imposed.
If not \code{NULL}, \code{bounds} is interpreted as:
\itemize{
\item lower bounds, if \code{cmp = .Primitive("<=")}, or
\item upper bounds, if \code{cmp = .Primitive(">=")}.
}

See also \code{total_cost}.}

\item{unit_costs}{(\code{numeric})\cr costs \eqn{c_1,\ldots,c_H} of surveying one
element in each stratum. Strictly positive values. May also be of length 1,
in which case the value is recycled to match the length of \code{bounds}.}

\item{cmp}{(\code{function})\cr a binary comparison operator used to check for
violations of \code{bounds}.
Must be either \code{.Primitive("<=")} (treating \code{bounds} as lower bounds and
invoking the LRNA algorithm) or \code{.Primitive(">=")} (treating \code{bounds} as
upper bounds and invoking the RNA algorithm).

The value of this argument has no effect if \code{bounds} is \code{NULL}.}

\item{details}{(\code{logical(1)})\cr should detailed information on stratum
assignments (either take-Neyman or take-bound), values of the set function
\eqn{s}, and the number of iterations be included in the output?}

\item{M}{(\code{numeric} or \code{NULL})\cr upper bounds \eqn{M_1,\ldots,M_H} optionally
imposed on sample sizes in strata. Set to \code{NULL} if no upper bounds are imposed.
Otherwise, it is required that \code{total_cost <= sum(unit_costs * M)}.}
}
\value{
A numeric vector of optimum sample allocations in strata. In the case
of \code{rna()} only, the return value may also be a \code{\link{list}} containing the
optimum allocations and strata assignments.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Functions implementing selected optimum allocation algorithms for solving
the optimum sample allocation problem, formulated as follows:

Minimize
\deqn{f(x_1,\ldots,x_H) = \sum_{h=1}^H \frac{A^2_h}{x_h}}
over \eqn{\mathbb R_+^H}, subject to
\deqn{\sum_{h=1}^H c_h x_h = c,}
and either
\deqn{x_h \leq M_h, \qquad h = 1,\ldots,H,}
or
\deqn{x_h \geq m_h, \qquad h = 1,\ldots,H,}
where
\eqn{c > 0,\, c_h > 0,\, A_h > 0,\, m_h > 0,\, M_h > 0,\, h = 1,\ldots,H},
are given numbers.

The following is a list of all available algorithms along with the functions
that implement them:
\itemize{
\item RNA - \code{rna()},
\item LRNA - \code{rna()},
\item SGA - \code{sga()},
\item SGAPLUS - \code{sgaplus()},
\item COMA - \code{coma()}.
}

See the documentation of a specific function for details.

The inequality constraints are optional. The user can choose whether and how
they are imposed in the optimization problem, depending on the chosen algorithm:
\itemize{
\item Lower bounds \eqn{m_1, \ldots, m_H} can be specified only for the LRNA
algorithm (by setting \code{cmp = .Primitive("<=")} for \code{rna()}).
\item Upper bounds \eqn{M_1, \ldots, M_H} are supported by all other algorithms.
\item Simultaneous constraints (both lower and upper bounds) are not supported
by these functions.
}

The costs \eqn{c_1, \ldots, c_H} of surveying one element in a stratum
can be specified by the user only for the RNA and LRNA algorithms.
For the remaining algorithms, these costs are fixed at 1, i.e.,
\eqn{c_h = 1,\, h = 1,\ldots,H}.
}
\details{
If no inequality constraints are imposed, the allocation is given by
the Neyman allocation:
\deqn{x_h = \frac{A_h}{\sqrt{c_h}} \frac{c}{\sum_{i=1}^H A_i \sqrt{c_i}},
  \qquad h = 1,\ldots,H.}

For the \emph{stratified \eqn{\pi}-estimator} of the population total under
\emph{stratified simple random sampling without replacement} design, the
parameters of the objective function \eqn{f} are
\deqn{A_h = N_h S_h, \qquad h = 1,\ldots,H,}
where \eqn{N_h} denotes the size of stratum \eqn{h} and \eqn{S_h} is the
standard deviation of the study variable in stratum \eqn{h}.
}
\section{Functions}{
\itemize{
\item \code{rna()}: Implements the Recursive Neyman Algorithm (RNA) and its counterpart,
the Lower Recursive Neyman Algorithm (LRNA), designed for the optimum
allocation problem with one-sided lower-bound constraints.
The RNA is described in \insertCite{WWW;textual}{stratallo}, whereas the LRNA
is introduced in \insertCite{WojciakLRNA;textual}{stratallo}.

\item \code{sga()}: The Stenger-Gabler (SGA) algorithm, as proposed by
\insertCite{SG;textual}{stratallo} and described in
\insertCite{WWW;textual}{stratallo}.
This algorithm solves the optimum allocation problem with one-sided
upper-bound constraints. It assumes unit costs are constant and equal to 1,
i.e., \eqn{c_h = 1,\, h = 1,\ldots,H}.

\item \code{sgaplus()}: A modified Stenger-Gabler-type algorithm, described in
\insertCite{WojciakMsc;textual}{stratallo}, implemented as the
Sequential Allocation (version 1) algorithm.
This algorithm solves the optimum allocation problem with one-sided
upper-bound constraints. It assumes unit costs are constant and equal to 1,
i.e., \eqn{c_h = 1,\, h = 1,\ldots,H}.

\item \code{coma()}: The Change of Monotonicity Algorithm (COMA), described in
\insertCite{WWW;textual}{stratallo}, solves the optimum allocation problem
with one-sided upper-bound constraints. It assumes unit costs are constant
and equal to 1, i.e., \eqn{c_h = 1,\, h = 1,\ldots,H}.

}}
\note{
These functions are optimized for internal use and should typically not
be called directly by users. Use \code{\link[=opt]{opt()}} or \code{\link[=optcost]{optcost()}} instead.
}
\examples{
A <- c(3000, 4000, 5000, 2000)
m <- c(50, 40, 10, 30) # lower bounds
M <- c(100, 90, 70, 80) # upper bounds

rna(total_cost = 190, A = A, bounds = M)

rna(total_cost = 190, A = A, bounds = m, cmp = .Primitive("<="))

rna(total_cost = 300, A = A, bounds = M)

sga(total_cost = 190, A = A, M = M)

sgaplus(total_cost = 190, A = A, M = M)

coma(total_cost = 190, A = A, M = M)

}
\references{
\insertRef{WojciakLRNA}{stratallo}

\insertRef{WWW}{stratallo}

\insertRef{WojciakMsc}{stratallo}

\insertRef{SG}{stratallo}

\insertRef{Sarndal}{stratallo}
}
\seealso{
\code{\link[=opt]{opt()}}, \code{\link[=optcost]{optcost()}}, \code{\link[=rnabox]{rnabox()}}.
}
