% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alloc_summary.R
\name{alloc_summary}
\alias{alloc_summary}
\title{Summarizing the Allocation}
\usage{
alloc_summary(x, A, m = NULL, M = NULL)
}
\arguments{
\item{x}{(\code{numeric})\cr sample allocations \eqn{x_1,\ldots,x_H}.}

\item{A}{(\code{numeric})\cr population constants \eqn{A_1,\ldots,A_H}.}

\item{m}{(\code{numeric} or \code{NULL})\cr optional lower bounds \eqn{m_1,\ldots,m_H}.}

\item{M}{(\code{numeric} or \code{NULL})\cr optional upper bounds \eqn{M_1,\ldots,M_H}.}
}
\value{
A \code{\link{data.frame}} with \eqn{H + 1} rows and up to seven variables,
where \eqn{H} is the number of strata.
The first \eqn{H} rows correspond to strata \eqn{h = 1,\ldots,H}, while
the last row contains column totals (where applicable).
The columns include:
\describe{
\item{A}{Population constant \eqn{A_h}.}
\item{m*}{Lower bound \eqn{m_h} (if provided).}
\item{M*}{Upper bound \eqn{M_h} (if provided).}
\item{allocation}{The optimized sample size \eqn{x_h}.}
\item{take_min*}{Boolean indicator: \eqn{x_h = m_h}.}
\item{take_max*}{Boolean indicator: \eqn{x_h = M_h}.}
\item{take_Neyman}{Boolean indicator: \eqn{m_h < x_h < M_h} (or
simply the internal Neyman allocation if no bounds were violated).}
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

A utility that returns a simple \code{\link{data.frame}} summarizing the allocation
returned by \code{\link[=opt]{opt()}} or \code{\link[=optcost]{optcost()}}.
}
\examples{
A <- c(3000, 4000, 5000, 2000)
m <- c(100, 90, 70, 80)
M <- c(200, 150, 300, 210)

xopt_1 <- opt(n = 400, A, m)
alloc_summary(xopt_1, A, m)

xopt_2 <- opt(n = 540, A, m, M)
alloc_summary(xopt_2, A, m, M)
}
\seealso{
\code{\link[=opt]{opt()}}, \code{\link[=optcost]{optcost()}}.
}
