% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\alias{pop10s_bounds_ucost}
\alias{pop507s_ucost}
\alias{pop969s_ucost}
\alias{pop2d4s}
\alias{pop9d278s}
\title{Datasets with Example Populations}
\format{
pop10s_bounds_ucost: Population with 10 strata, lower and upper bounds
on sample sizes, and associated surveying costs.
A matrix with 10 rows and 5 variables:
\describe{
\item{N}{stratum size}
\item{S}{standard deviation of study variable in the stratum.}
\item{m}{lower bound for sample size in the stratum.}
\item{M}{upper bound for sample size in the stratum.}
\item{unit_cost}{cost of surveying one element in the stratum.}
}

pop507s_ucost: Population with 507 strata and associated surveying costs.
A matrix with 507 rows and 3 columns:
\describe{
\item{N}{stratum size.}
\item{S}{standard deviation of study variable in the stratum.}
\item{unit_cost}{cost of surveying one element in the stratum.}
}

pop969s_ucost: Population with 969 strata and associated surveying costs.
A matrix with 969 rows and 3 columns:
\describe{
\item{N}{stratum size.}
\item{S}{standard deviation of study variable in the stratum.}
\item{unit_cost}{cost of surveying one element in the stratum.}
}

pop2d4s: Population with 2 domains and 4 strata.
A list with the following elements:
\describe{
\item{H_counts}{strata counts in each domain.}
\item{N}{stratum sizes.}
\item{S}{standard deviations of study variable in strata.}
\item{total}{totals in domains, i.e., the sum of the study variable values
for population elements in each domain.}
\item{kappa}{priority weights for domains.}
\item{rho}{\code{total * sqrt(kappa)}.}
\item{rho2}{\code{total^2 * kappa}.}
\item{n_max}{See \code{\link[=dca_nmax]{dca_nmax()}} or \code{\link[=dca]{dca()}}.}
}

pop9d278s: Population with 9 domains and 278 strata.
A list with the following elements:
\describe{
\item{H_counts}{strata counts in each domain.}
\item{N}{stratum sizes.}
\item{S}{standard deviations of study variable in strata.}
\item{total}{totals in domains, i.e., the sum of the study variable values
for population elements in each domain.}
\item{kappa}{priority weights for domains.}
\item{rho}{\code{total * sqrt(kappa)}.}
\item{rho2}{\code{total^2 * kappa}.}
\item{n_max}{See \code{\link[=dca_nmax]{dca_nmax()}} or \code{\link[=dca]{dca()}}.}
}
}
\usage{
pop10s_bounds_ucost

pop507s_ucost

pop969s_ucost

pop2d4s

pop9d278s
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Example datasets containing artificial populations for testing and
demonstrating optimum sample allocation algorithms.
}
\keyword{datasets}
