% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_dca.R
\name{dca}
\alias{dca}
\alias{dca0}
\alias{dca_nmax}
\title{Domain-Controlled Allocation (DCA) Algorithm}
\usage{
dca0(n, H_counts, N, S, rho, rho2, details = FALSE)

dca(n, H_counts, N, S, rho, rho2, U = NULL, details = FALSE)

dca_nmax(H_counts, N, S)
}
\arguments{
\item{n}{(\code{integerish(1)})\cr total sample size \eqn{n}. Must satisfy \verb{0 < n <= sum(N)}.}

\item{H_counts}{(\code{integerish})\cr strata counts in each domain.}

\item{N}{(\code{integerish})\cr strata sizes \eqn{(N_{d,h},\, (d,h) \in \mathcal H)}.}

\item{S}{(\code{numeric})\cr standard deviations \eqn{(S_{d,h},\, (d,h) \in \mathcal H)}
of surveyed variable in strata.}

\item{rho}{(\code{numeric})\cr parameters \eqn{(\rho_d,\, d \in \mathcal D)} of the
optimization problem.}

\item{rho2}{(\code{numeric})\cr the square of \code{rho} (\code{rho^2}), provided to reduce
potential loss of precision due to finite-precision arithmetic.}

\item{details}{(\code{logical(1)})\cr whether to produce detailed debug output.}

\item{U}{(\code{integerish} or \code{NULL})\cr a vector of indices identifying the
\emph{take-max} strata, i.e., the strata \eqn{(d,h)} for which the
allocation is fixed to \eqn{x_{d,h} = N_{d,h}}.
The indices refer to the positions of strata in the set \eqn{\mathcal H},
in the same order as in the input vectors (\code{N}, \code{S}, etc.).

For example, if \eqn{\mathcal H = \{(1,1),\, (2,1)\}} and
stratum \eqn{(2,1)} is a take-max stratum, then \code{U = 2}.

If \code{U} contains all strata from a domain, the dimension of the
D matrix is reduced accordingly.

\code{U} must satisfy one of the following conditions:
\itemize{
\item \code{n > sum(N[U])},
\item \code{n = sum(N[U])} and \code{n = sum(N)}.
}}
}
\value{
If \code{details = FALSE}, the optimal \eqn{\boldsymbol x^*} is returned.
Otherwise, a list is returned containing the optimal \eqn{\boldsymbol x^*}
(element named \code{x}) along with other internal details of this algorithm.
In particular, the \code{lambda} element of the list corresponds to the optimal
\eqn{T^*}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Functions implementing the Domain-Controlled Allocation (DCA)
algorithm described in \insertCite{Wesolowski;textual}{stratallo} and
\insertCite{WojciakPhD;textual}{stratallo}.
The algorithm solves the following
optimum allocation problem, formulated in mathematical optimization terms:

Minimize
\deqn{f(T,\, \boldsymbol x) = T}
over \eqn{\mathbb R \times \mathbb R_+^{\lvert \mathcal H \rvert}},
subject to
\deqn{\sum_{(d,h) \in \mathcal H} x_{d,h} = n,}
\deqn{\sum_{h \in \mathcal H_d} (\frac{1}{x_{d,h}} - \frac{1}{N_{d,h}}) \frac{N_{d,h}^2 S_{d,h}^2}{\rho_d^2} = T, \qquad d \in \mathcal D,}
where:
\describe{
\item{\eqn{(T,\, \boldsymbol x) = (T,\, (x_{d,h},\, (d,h) \in \mathcal H))}}{
the optimization variable,
}
\item{\eqn{\mathcal H \subset \mathbb N^2}}{
the set of domain-stratum indices,
}
\item{\eqn{\mathcal D := \{d \in \mathbb N \colon\; \exists h,\, (d,h) \in \mathcal H\}}}{
the set of domain indices,
}
\item{\eqn{\mathcal H_d := \{h \in \mathbb N \colon\; (d,h) \in \mathcal H\}}}{
the set of strata indices in domain \eqn{d},
}
\item{\eqn{N_{d,h} > 0}}{size of stratum \eqn{(d,h)},}
\item{\eqn{S_{d,h} > 0}}{standard deviation of the study variable in stratum \eqn{(d,h)},}
\item{\eqn{\rho_d := t_d\, \sqrt{\kappa_d}}}{
where \eqn{t_d} denotes the total in domain \eqn{d}, i.e., the sum of the
values of the study variable for population elements in domain \eqn{d},
and \eqn{\kappa_d} is a priority weight for domain \eqn{d},
}
\item{\eqn{n \in (0,\, \sum_{(d,h) \in \mathcal H} N_{d,h}]}}{total sample size.}
}
}
\details{
For \eqn{n \in (0,\, n_{max})}, the optimal value satisfies \eqn{T^* > 0},
where
\deqn{
    n_{max} := \sum_{d \in \mathcal D}
    \frac{\bigl( \sum_{h \in \mathcal H_d} N_{d,h} S_{d,h} \bigr)^2}{\sum_{h \in \mathcal H_d} N_{d,h} S_{d,h}^2}.
  }
See Proposition 2.1 in \insertCite{Wesolowski;textual}{stratallo} or
\insertCite{WojciakPhD;textual}{stratallo} for details.
The value \eqn{n_{max}} is less than or equal to \code{sum(N)} and can be
computed with \code{\link[=dca_nmax]{dca_nmax()}}.
}
\section{Functions}{
\itemize{
\item \code{dca0()}: Domain-Controlled Allocation algorithm by
\insertCite{Wesolowski;textual}{stratallo}

\item \code{dca()}: Domain-Controlled Allocation algorithm by
\insertCite{Wesolowski;textual}{stratallo}, optionally using a set of
take-max strata as described in \insertCite{WojciakPhD;textual}{stratallo}.

\item \code{dca_nmax()}: Computes the maximum total sample size \eqn{n_{max}} such that the
optimization problem solved by the Domain-Controlled Allocation (DCA)
algorithm admits a strictly positive optimal value \eqn{T^*}.

}}
\note{
These functions are optimized for internal use and should typically not
be called directly by users. They are designed to handle a large number of
invocations, specifically recursive calls from \code{rdca()}, and, as a result,
parameter assertions are minimal.
}
\examples{
# Two domains with 1 and 3 strata, respectively,
# that is, H = {(1,1), (2,1), (2,2), (2,3)}.
H_counts <- c(1, 3)
N <- c(140, 110, 135, 190) # (N_{1,1}, N_{2,1}, N_{2,2}, N_{2,3})
S <- sqrt(c(180, 20, 5, 4)) # (S_{1,1}, S_{2,1}, S_{2,2}, S_{2,3})
total <- c(2, 3)
kappa <- c(0.4, 0.6)
rho <- total * sqrt(kappa) # (rho_1, rho_2)
rho2 <- total^2 * kappa
sum(N) # 575
n_max <- dca_nmax(H_counts, N, S) # 519.0416

n <- floor(n_max) - 1

dca0(n, H_counts, N, S, rho, rho2)
x0 <- dca0(n, H_counts, N, S, rho, rho2, details = TRUE)
x0$x
x0$lambda
x0$k
x0$v
x0$s

n <- ceiling(n_max) + 1
x0 <- dca0(n, H_counts, N, S, rho, rho2, details = TRUE)
x0$x
x0$lambda

n <- floor(n_max) - 1

x1 <- dca(n, H_counts, N, S, rho, rho2, details = TRUE)
x1$x
x1$x_Uc
x1$lambda
x1$s

dca(n, H_counts, N, S, rho, rho2, U = 1)
x2 <- dca(n, H_counts, N, S, rho, rho2, U = 1, details = TRUE)
x2$x
x2$x_Uc
x2$lambda
x2$s

}
\references{
\insertRef{WojciakPhD}{stratallo}

\insertRef{Wesolowski}{stratallo}

\insertRef{WJWR2017}{stratallo}
}
\seealso{
\code{\link[=rdca]{rdca()}}
}
