% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_dca.R
\name{dca_M}
\alias{dca_M}
\title{DCA Algorithm for Upper-Bound Constrained Allocations (M <= N)}
\usage{
dca_M(n, H_counts, N, S, rho, rho2, M = N, U = NULL)
}
\arguments{
\item{n}{(\code{integerish(1)})\cr total sample size \eqn{n}. Must satisfy \verb{0 < n <= sum(N)}.}

\item{H_counts}{(\code{integerish})\cr strata counts in each domain.}

\item{N}{(\code{integerish})\cr strata sizes \eqn{(N_{d,h},\, (d,h) \in \mathcal H)}.}

\item{S}{(\code{numeric})\cr standard deviations \eqn{(S_{d,h},\, (d,h) \in \mathcal H)}
of surveyed variable in strata.}

\item{rho}{(\code{numeric})\cr parameters \eqn{(\rho_d,\, d \in \mathcal D)} of the
optimization problem.}

\item{rho2}{(\code{numeric})\cr the square of \code{rho} (\code{rho^2}), provided to reduce
potential loss of precision due to finite-precision arithmetic.}

\item{M}{(\code{integerish})\cr upper bounds on sample sizes in strata.
Defaults to \code{N}.}

\item{U}{(\code{integerish} or \code{NULL})\cr a vector of indices identifying the
\emph{take-max} strata, i.e., the strata \eqn{(d,h)} for which the
allocation is fixed to \eqn{x_{d,h} = N_{d,h}}.
The indices refer to the positions of strata in the set \eqn{\mathcal H},
in the same order as in the input vectors (\code{N}, \code{S}, etc.).

For example, if \eqn{\mathcal H = \{(1,1),\, (2,1)\}} and
stratum \eqn{(2,1)} is a take-max stratum, then \code{U = 2}.

If \code{U} contains all strata from a domain, the dimension of the
D matrix is reduced accordingly.

\code{U} must satisfy one of the following conditions:
\itemize{
\item \code{n > sum(N[U])},
\item \code{n = sum(N[U])} and \code{n = sum(N)}.
}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Prototype (under testing).
}
\examples{
H_counts <- c(5, 2)
H_names <- rep(seq_along(H_counts), times = H_counts)
S <- c(154, 178, 134, 213, 124, 102, 12)
N <- c(100, 100, 100, 100, 100, 100, 100)
M <- c(80, 90, 70, 40, 10, 90, 100)
names(M) <- names(N) <- H_names
total <- c(13, 2)
kappa <- c(0.8, 0.2)
n <- 150

# experimental function (not exported) – examples skipped
\dontrun{
dca_M(n, H_counts, N, S, total, kappa, M = M, U = 5)
#         1         1         1         1         1         2         2
# 12.754880 14.742653 11.098402 17.641490 10.000000 74.945462  8.817113
}

}
