% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dopt.R
\name{dopt}
\alias{dopt}
\title{Multi-Domain Optimum Sample Allocation with Controlled-Precision under
Upper-Bound Constraints}
\usage{
dopt(n, H_counts, N, S, total, kappa, return_T = FALSE)
}
\arguments{
\item{n}{(\code{integerish(1)})\cr total sample size \eqn{n}. Must satisfy \verb{0 < n <= sum(N)}.}

\item{H_counts}{(\code{integerish})\cr strata counts in each domain.}

\item{N}{(\code{integerish})\cr strata sizes \eqn{(N_{d,h},\, (d,h) \in \mathcal H)}.}

\item{S}{(\code{numeric})\cr standard deviations \eqn{(S_{d,h},\, (d,h) \in \mathcal H)}
of surveyed variable in strata.}

\item{total}{(\code{numeric})\cr vector of domain totals,
\eqn{t_d,\, d \in \mathcal D}, i.e., the sum of the study variable over all
population elements in each domain.}

\item{kappa}{(\code{numeric})\cr vector of priority weights for the domains,
\eqn{\kappa_d,\, d \in \mathcal D}.}

\item{return_T}{(\code{logical(1)})\cr
If \code{TRUE}, the function returns a list containing the optimal allocation
and the optimal value of the objective function \eqn{T}. If \code{FALSE}
(default), only the optimal allocation vector is returned.}
}
\value{
If \code{return_T = FALSE} (default), a numeric vector containing the optimal
sample allocations \eqn{x_{d,h}} for each stratum \eqn{(d,h) \in \mathcal H}.

If \code{return_T = TRUE}, a list with components:
\describe{
\item{xopt}{numeric vector of optimal sample allocations.}
\item{Topt}{optimal value of the objective function \eqn{T}.}
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Computes the optimum allocation for the following multi-domain optimum
allocation problem, formulated in mathematical optimization terms:

Minimize
\deqn{f(T,\, \boldsymbol x) = T}
over \eqn{\mathbb R \times \mathbb R_+^{\lvert \mathcal H \rvert}},
subject to
\deqn{\sum_{(d,h) \in \mathcal H} x_{d,h} = n,}
\deqn{\sum_{h \in \mathcal H_d} (\frac{1}{x_{d,h}} - \frac{1}{N_{d,h}}) \frac{N_{d,h}^2 S_{d,h}^2}{\rho_d^2} = T, \qquad d \in \mathcal D,}
\deqn{x_{d,h} \leq N_{d,h}, \qquad (d,h) \in \mathcal H,}
where:
\describe{
\item{\eqn{(T,\, \boldsymbol x) = (T,\, (x_{d,h},\, (d,h) \in \mathcal H))}}{
the optimization variable,
}
\item{\eqn{\mathcal H \subset \mathbb N^2}}{
the set of domain-stratum indices,
}
\item{\eqn{\mathcal D := \{d \in \mathbb N \colon\; \exists h,\, (d,h) \in \mathcal H\}}}{
the set of domain indices,
}
\item{\eqn{\mathcal H_d := \{h \in \mathbb N \colon\; (d,h) \in \mathcal H\}}}{
the set of strata indices in domain \eqn{d},
}
\item{\eqn{N_{d,h} > 0}}{size of stratum \eqn{(d,h)},}
\item{\eqn{S_{d,h} > 0}}{standard deviation of the study variable in stratum \eqn{(d,h)},}
\item{\eqn{\rho_d := t_d\, \sqrt{\kappa_d}}}{
where \eqn{t_d} denotes the total in domain \eqn{d}, i.e., the sum of the
values of the study variable for population elements in domain \eqn{d},
and \eqn{\kappa_d} is a priority weight for domain \eqn{d},
}
\item{\eqn{n \in (0,\, \sum_{(d,h) \in \mathcal H} N_{d,h}]}}{total sample size.}
}
}
\details{
The \code{dopt()} function uses the RDCA algorithm implemented in \code{\link[=rdca]{rdca()}}.
}
\examples{

# Three domains with 2, 2, and 3 strata, respectively,
# that is, H = {(1,1), (1,2), (2,1), (2,2), (3,1), (3,2), (3,3)}.
H_counts <- c(2, 2, 3)
# (N_{1,1}, N_{1,2}, N_{2,1}, N_{2,2}, N_{3,1}, N_{3,2}, N_{3,3})
N <- c(140, 110, 135, 190, 200, 40, 70)
# (S_{1,1}, S_{1,2}, S_{2,1}, S_{2,2}, S_{3,1}, S_{3,2}, S_{3,3})
S <- c(180, 20, 5, 4, 35, 9, 40)
total <- c(2, 3, 5)
kappa <- c(0.5, 0.2, 0.3)
n <- 828

# Optimum allocation.
dopt(n, H_counts, N, S, total, kappa)

# Example population with 9 domains and 278 strata
p <- pop9d278s
sum(p$N)
n <- 5000
x <- dopt(n, p$H_counts, p$N, p$S, p$total, p$kappa, return_T = TRUE)
x
all(x$xopt <= p$N)
sum(x$xopt)

}
\references{
\insertRef{WojciakPhD}{stratallo}
}
\seealso{
\code{\link[=rdca]{rdca()}}, \code{\link[=dca]{dca()}}, \code{\link[=dca_nmax]{dca_nmax()}}, \code{\link[=opt]{opt()}}, \code{\link[=optcost]{optcost()}}.
}
