% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt.R
\name{opt}
\alias{opt}
\title{Optimum Sample Allocation in Stratified Sampling}
\usage{
opt(n, A, m = NULL, M = NULL, M_algorithm = "rna")
}
\arguments{
\item{n}{(\code{integerish(1)})\cr total sample size.
Must satisfy \code{n > 0}.
Additionally:
\itemize{
\item If \code{bounds_inner} is not \code{NULL}, then
\code{n >= sum(bounds_inner)} when \code{bounds_inner} are treated as lower bounds, or
\code{n <= sum(bounds_inner)} when treated as upper bounds.
\item If \code{bounds_outer} is not \code{NULL}, then
\code{n >= sum(bounds_outer)} when \code{bounds_outer} are treated as lower bounds, or
\code{n <= sum(bounds_outer)} when treated as upper bounds.
}}

\item{A}{(\code{numeric})\cr population constants \eqn{A_1,\ldots,A_H}.
All values must be strictly positive.}

\item{m}{(\code{numeric} or \code{NULL})\cr
optional lower bounds \eqn{m_1,\ldots,m_H} for the stratum sample sizes.
If no lower bounds are desired, set \code{m = NULL}. If \code{M} is not \code{NULL}, it is
required that \eqn{m_h < M_h} for all strata.}

\item{M}{(\code{numeric} or \code{NULL})\cr
optional upper bounds \eqn{M_1,\ldots,M_H} for the stratum sample sizes.
If no upper bounds are desired, set \code{M = NULL}. If \code{m} is not \code{NULL}, it is
required that \eqn{m_h < M_h} for all strata.}

\item{M_algorithm}{(\code{string})\cr
Name of the algorithm to use for computing the sample allocation when only
upper-bound constraints are imposed.
Must be one of \code{"rna"} (default), \code{"sga"}, \code{"sgaplus"}, or \code{"coma"}.
This parameter is used only when \eqn{H > 1} and \code{n < sum(M)}.}
}
\value{
A numeric vector of the optimal sample allocations for each stratum.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Computes the optimum allocation for the following optimum allocation problem,
formulated in mathematical optimization terms:

Minimize
\deqn{f(x_1,\ldots,x_H) = \sum_{h=1}^H \frac{A^2_h}{x_h}}
over \eqn{\mathbb R_+^H}, subject to
\deqn{\sum_{h=1}^H x_h = n,}
\deqn{m_h \leq x_h \leq M_h, \qquad h = 1,\ldots,H,}
where \eqn{n > 0,\, A_h > 0,\, m_h > 0,\, M_h > 0}, such that
\eqn{m_h < M_h,\, h = 1,\ldots,H}, and
\eqn{\sum_{h=1}^H m_h \leq n \leq \sum_{h=1}^H M_h}, are given numbers.
Inequality constraints are optional and may be omitted.

Inequality constraints are optional, and the user can choose whether and how
they are applied to the optimization problem. This is controlled using the
\code{m} and \code{M} arguments as follows:
\itemize{
\item \strong{No inequality constraints:} both \code{m} and \code{M} must be \code{NULL} (default).
\item \strong{Lower bounds only (\eqn{m_1,\, \ldots,\, m_H}):} specify \code{m}, and
set \code{M = NULL}.
\item \strong{Upper bounds only (\eqn{M_1,\, \ldots,\, M_H}):} specify \code{M}, and
set \code{m = NULL}.
\item \strong{Box constraints (\eqn{m_h, M_h,\, h = 1,\ldots,H}):} specify both \code{m}
and \code{M}.
}
}
\details{
The \code{opt()} function uses different allocation algorithms depending on which
inequality constraints are applied. Each algorithm is implemented in a
separate R function, which is generally \strong{not intended to be called directly}
by the end user. The algorithms are:
\itemize{
\item \strong{Lower bounds only (\eqn{m_1,\, \ldots,\, m_H}):}
\itemize{
\item \code{LRNA} - \code{\link[=rna]{rna()}}
}
\item \strong{Upper bounds only (\eqn{M_1,\, \ldots,\, M_H}):}
\itemize{
\item \code{RNA} - \code{\link[=rna]{rna()}}
\item \code{SGA} - \code{\link[=sga]{sga()}}
\item \code{SGAPLUS} - \code{\link[=sgaplus]{sgaplus()}}
\item \code{COMA} - \code{\link[=coma]{coma()}}
}
\item \strong{Box constraints (\eqn{m_h, M_h,\, h = 1,\ldots,H}):}
\itemize{
\item \code{RNABOX} - \code{\link[=rnabox]{rnabox()}}
}
}

See the documentation of each specific function for more details about the
corresponding algorithm.
}
\note{
If no inequality constraints are applied, the allocation follows the
Neyman allocation:
\deqn{x_h = A_h \frac{n}{\sum_{i=1}^H A_i}, \quad h = 1,\ldots,H.}

For a \strong{stratified \eqn{\pi} estimator} of the population total using
stratified simple random sampling without replacement design, the objective
function parameters \eqn{A_h} are:
\deqn{A_h = N_h S_h, \quad h = 1,\ldots,H,}
where \eqn{N_h} is the size of stratum \eqn{h} and \eqn{S_h} is the
standard deviation of the study variable in stratum \eqn{h}.
}
\examples{
A <- c(3000, 4000, 5000, 2000)
m <- c(100, 90, 70, 50)
M <- c(300, 400, 200, 90)

# One-sided lower bounds.
opt(n = 340, A = A, m = m)
opt(n = 400, A = A, m = m)
opt(n = 700, A = A, m = m)

# One-sided upper bounds.
opt(n = 190, A = A, M = M)
opt(n = 700, A = A, M = M)

# Box-constraints.
opt(n = 340, A = A, m = m, M = M)
opt(n = 500, A = A, m = m, M = M)
x <- opt(n = 800, A = A, m = m, M = M)
x

# Variance corresponding to the allocation x.
var_st(x = x, A = A, A0 = 45000)

# Execution-time comparison of different algorithms using the microbenchmark package.
\dontrun{
N <- pop969s_ucost[, "N"]
S <- pop969s_ucost[, "S"]
A <- N * S
nfrac <- c(0.005, seq(0.05, 0.95, 0.05))
n <- setNames(as.integer(nfrac * sum(N)), nfrac)
lapply(
  n,
  function(ni) {
    microbenchmark::microbenchmark(
      RNA = opt(ni, A, M = N, M_algorithm = "rna"),
      SGA = opt(ni, A, M = N, M_algorithm = "sga"),
      SGAPLUS = opt(ni, A, M = N, M_algorithm = "sgaplus"),
      COMA = opt(ni, A, M = N, M_algorithm = "coma"),
      times = 200,
      unit = "us"
    )
  }
)
}
}
\references{
\insertRef{Sarndal}{stratallo}
}
\seealso{
\code{\link[=optcost]{optcost()}}, \code{\link[=rna]{rna()}}, \code{\link[=sga]{sga()}}, \code{\link[=sgaplus]{sgaplus()}}, \code{\link[=coma]{coma()}}, \code{\link[=rnabox]{rnabox()}}.
}
