% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt.R
\name{optcost}
\alias{optcost}
\title{Minimum-Cost Allocation in Stratified Sampling}
\usage{
optcost(V, A, A0, M = NULL, unit_costs = 1)
}
\arguments{
\item{V}{(\code{number})\cr parameter \eqn{V} in the variance constraint.
If upper bounds are imposed (\code{M} is not \code{NULL}), it must satisfy
\code{V > sum(A^2/M) - A_0}. Otherwise, \code{V > 0}.}

\item{A}{(\code{numeric})\cr population constants \eqn{A_1,\ldots,A_H}.
All values must be strictly positive.}

\item{A0}{(\code{number})\cr population constant \eqn{A_0}.}

\item{M}{(\code{numeric} or \code{NULL})\cr optional upper bounds \eqn{M_1,\ldots,M_H}
on the stratum sample sizes.
If no upper bounds are imposed, set \code{M = NULL}.}

\item{unit_costs}{(\code{numeric})\cr costs \eqn{c_1,\ldots,c_H} of surveying one
element in each stratum. Strictly positive values. May also be of length 1,
in which case the value is recycled to match the length of \code{bounds}.}
}
\value{
A numeric vector containing the optimal sample allocation for each stratum.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Computes stratum sample sizes that minimize the total survey cost for a
given target variance of a stratified estimator, optionally subject to
one-sided upper bounds on the stratum sample sizes.
Specifically, the function solves the following optimization problem:

Minimize
\deqn{c(x_1,\ldots,x_H) = \sum_{h=1}^H c_h x_h}
over \eqn{\mathbb R_+^H}, subject to
\deqn{\sum_{h=1}^H \frac{A^2_h}{x_h} - A_0 = V,}
\deqn{x_h \leq M_h, \qquad h = 1,\ldots,H,}
where \eqn{A_0,\, A_h > 0,\, c_h > 0,\, M_h > 0,\, h = 1,\ldots,H},
and \eqn{V > \sum_{h=1}^H \frac{A^2_h}{M_h} - A_0}, are given numbers.

The upper-bound constraints \eqn{x_h \leq M_h} are optional. If they are not
imposed, it is only required that \eqn{V > 0}.
}
\details{
The allocation is computed using the \strong{LRNA algorithm}, described in
\insertCite{WojciakLRNA;textual}{stratallo}.

The solution is valid for stratified sampling designs in which the variance
\eqn{V_{st}} of the stratified estimator can be expressed as
\deqn{V_{st} = \sum_{h=1}^H \frac{A^2_h}{x_h} - A_0,}
where \eqn{H} is the number of strata, \eqn{x_1,\ldots,x_H} are the stratum
sample sizes, and \eqn{A_0,\, A_h > 0} do not depend on \eqn{x_h}.
}
\note{
For the \emph{stratified \eqn{\pi}-estimator} of the population total under
\emph{stratified simple random sampling without replacement} design, the
parameters take the form
\deqn{A_h = N_h S_h, \qquad h = 1,\ldots,H,}
\deqn{A_0 = \sum_{h=1}^H N_h S_h^2,}
where \eqn{N_h} is the size of stratum \eqn{h} and \eqn{S_h} is the
standard deviation of the study variable in stratum \eqn{h}.
}
\examples{
A <- c(3000, 4000, 5000, 2000)
M <- c(100, 90, 70, 80)
x <- optcost(1017579, A = A, A0 = 579, M = M)
x
}
\references{
\insertRef{WojciakLRNA}{stratallo}
}
\seealso{
\code{\link[=rna]{rna()}}, \code{\link[=opt]{opt()}}.
}
