% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop10_mM}
\alias{pop10_mM}
\title{Example Population with 10 Strata and Lower and Upper Bounds}
\format{
A matrix with 10 rows and 5 variables:
\describe{
\item{N}{stratum size}
\item{S}{standard deviation of study variable in stratum}
\item{m}{lower bound for sample size in stratum}
\item{M}{upper bound for sample size in stratum}
\item{unit_cost}{cost of surveying one element in stratum}
}
}
\usage{
pop10_mM
}
\description{
A dataset containing the artificial population with 10 strata.
Additionally, the lower and upper bounds for samples in strata are
specified.
}
\keyword{datasets}
