% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop969}
\alias{pop969}
\title{Example Population with 969 Strata}
\format{
A matrix with 969 rows and 3 variables:
\describe{
\item{N}{stratum size}
\item{S}{standard deviation of study variable in stratum}
\item{unit_cost}{cost of surveying one element in stratum}
}
}
\usage{
pop969
}
\description{
A dataset containing the artificial population with 969 strata.
}
\keyword{datasets}
