% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_rdca.R
\name{rdca}
\alias{rdca}
\title{Recursive Domain-Controlled Allocation (RDCA) Algorithm}
\usage{
rdca(n, H_counts, N, S, rho, rho2 = rho^2, U = NULL, J = NULL)
}
\arguments{
\item{n}{(\code{integerish(1)})\cr total sample size \eqn{n}. Must satisfy \verb{0 < n <= sum(N)}.}

\item{H_counts}{(\code{integerish})\cr strata counts in each domain.}

\item{N}{(\code{integerish})\cr strata sizes \eqn{(N_{d,h},\, (d,h) \in \mathcal H)}.}

\item{S}{(\code{numeric})\cr standard deviations \eqn{(S_{d,h},\, (d,h) \in \mathcal H)}
of surveyed variable in strata.}

\item{rho}{(\code{numeric})\cr parameters \eqn{(\rho_d,\, d \in \mathcal D)} of the
optimization problem.}

\item{rho2}{(\code{numeric})\cr the square of \code{rho} (\code{rho^2}), provided to reduce
potential loss of precision due to finite-precision arithmetic.}

\item{U}{(\code{integerish} or \code{NULL})\cr a vector of indices identifying the
\emph{take-max} strata, i.e., the strata \eqn{(d,h)} for which the
allocation is fixed to \eqn{x_{d,h} = N_{d,h}}.
The indices refer to the positions of strata in the set \eqn{\mathcal H},
in the same order as in the input vectors (\code{N}, \code{S}, etc.).

For example, if \eqn{\mathcal H = \{(1,1),\, (2,1)\}} and
stratum \eqn{(2,1)} is a take-max stratum, then \code{U = 2}.

If \code{U} contains all strata from a domain, the dimension of the
D matrix is reduced accordingly.

\code{U} must satisfy one of the following conditions:
\itemize{
\item \code{n > sum(N[U])},
\item \code{n = sum(N[U])} and \code{n = sum(N)}.
}}

\item{J}{(\code{integerish} or \code{NULL})\cr vector of domain indices, subset of
\eqn{\mathcal D}.
Specifies the domains for which allocated sample sizes must not exceed
the corresponding strata sizes \code{N}.
For domains not included in \code{J}, allocations may exceed strata sizes.
Must not be empty.
If \code{J} is \code{NULL}, it is treated as containing all domains.
\code{U} must not contain any strata from domains included in \code{J}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Implements the Recursive Domain-Controlled Allocation (RDCA) algorithm
described in \insertCite{WojciakPhD;textual}{stratallo}.
The algorithm solves the following optimum allocation problem,
formulated in mathematical optimization terms:

Minimize
\deqn{f(T,\, \boldsymbol x) = T}
over \eqn{\mathbb R \times \mathbb R_+^{\lvert \mathcal H \rvert}},
subject to
\deqn{\sum_{(d,h) \in \mathcal H} x_{d,h} = n,}
\deqn{\sum_{h \in \mathcal H_d} (\frac{1}{x_{d,h}} - \frac{1}{N_{d,h}}) \frac{N_{d,h}^2 S_{d,h}^2}{\rho_d^2} = T, \qquad d \in \mathcal D,}
\deqn{x_{d,h} \leq N_{d,h}, \qquad (d,h) \in \mathcal H,}
where:
\describe{
\item{\eqn{(T,\, \boldsymbol x) = (T,\, (x_{d,h},\, (d,h) \in \mathcal H))}}{
the optimization variable,
}
\item{\eqn{\mathcal H \subset \mathbb N^2}}{
the set of domain-stratum indices,
}
\item{\eqn{\mathcal D := \{d \in \mathbb N \colon\; \exists h,\, (d,h) \in \mathcal H\}}}{
the set of domain indices,
}
\item{\eqn{\mathcal H_d := \{h \in \mathbb N \colon\; (d,h) \in \mathcal H\}}}{
the set of strata indices in domain \eqn{d},
}
\item{\eqn{N_{d,h} > 0}}{size of stratum \eqn{(d,h)},}
\item{\eqn{S_{d,h} > 0}}{standard deviation of the study variable in stratum \eqn{(d,h)},}
\item{\eqn{\rho_d := t_d\, \sqrt{\kappa_d}}}{
where \eqn{t_d} denotes the total in domain \eqn{d}, i.e., the sum of the
values of the study variable for population elements in domain \eqn{d},
and \eqn{\kappa_d} is a priority weight for domain \eqn{d},
}
\item{\eqn{n \in (0,\, \sum_{(d,h) \in \mathcal H} N_{d,h}]}}{total sample size.}
}
}
\details{
The upper-bound constraints \eqn{x_{d,h} \le N_{d,h}} are guaranteed
to be preserved only if domain \eqn{d} is in \code{J}.
The parameter \code{J} is used in the recursion.
The specified optimization problem is solved when \code{J = NULL}, i.e., when
\code{J} contains all domains.
}
\note{
This function is optimized for internal use and should typically not be
called directly by users.
It is designed to handle a large number of invocations, specifically
recursive invocations of \code{rdca()}, and, as a result, parameter assertions
are minimal.
}
\examples{
# Three domains with 2, 2, and 3 strata, respectively,
# that is, H = {(1,1), (1,2), (2,1), (2,2), (3,1), (3,2), (3,3)}.
H_counts <- c(2, 2, 3)
# (N_{1,1}, N_{1,2}, N_{2,1}, N_{2,2}, N_{3,1}, N_{3,2}, N_{3,3})
N <- c(140, 110, 135, 190, 200, 40, 70)
# (S_{1,1}, S_{1,2}, S_{2,1}, S_{2,2}, S_{3,1}, S_{3,2}, S_{3,3})
S <- c(180, 20, 5, 4, 35, 9, 40)
total <- c(2, 3, 5)
kappa <- c(0.5, 0.2, 0.3)
rho <- total * sqrt(kappa) # (rho_1, rho_2, rho_3)
rho2 <- total^2 * kappa
sum(N)
n <- 828

# Optimum allocation.
rdca(n, H_counts, N, S, rho, rho2)

# Upper bounds enforced only for domain 1.
rdca(n, H_counts, N, S, rho, rho2, J = 1)

}
\references{
\insertRef{WojciakPhD}{stratallo}
}
\seealso{
\code{\link[=dca]{dca()}}
}
