% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_rdca.R
\name{rdca_iter}
\alias{rdca_iter}
\title{Iterative RDCA Implementation}
\usage{
rdca_iter(n, H_counts, N, S, rho, rho2 = NULL, ref_domain = 1L)
}
\arguments{
\item{n}{(\code{integerish(1)})\cr total sample size \eqn{n}. Must satisfy \verb{0 < n <= sum(N)}.}

\item{H_counts}{(\code{integerish})\cr strata counts in each domain.}

\item{N}{(\code{integerish})\cr strata sizes \eqn{(N_{d,h},\, (d,h) \in \mathcal H)}.}

\item{S}{(\code{numeric})\cr standard deviations \eqn{(S_{d,h},\, (d,h) \in \mathcal H)}
of surveyed variable in strata.}

\item{rho}{(\code{numeric})\cr parameters \eqn{(\rho_d,\, d \in \mathcal D)} of the
optimization problem.}

\item{rho2}{(\code{numeric})\cr the square of \code{rho} (\code{rho^2}), provided to reduce
potential loss of precision due to finite-precision arithmetic.}

\item{ref_domain}{(\code{integerish(1)})\cr reference domain (denoted by j in the
thesis).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Iterative implementation of the Recursive Domain-Controlled Allocation (RDCA)
algorithm. Not tested.
}
\examples{
H_counts <- c(2, 2, 3)
N <- c(140, 110, 135, 190, 200, 40, 70)
S <- sqrt(c(180, 20, 5, 4, 35, 9, 40))
total <- c(2, 3, 5)
kappa <- c(0.5, 0.2, 0.3)
rho <- total * sqrt(kappa)
(n <- dca_nmax(H_counts, N, S) - 1)

# experimental function (not exported) – examples skipped
\dontrun{
rdca_iter(n, H_counts, N, S, rho)
# 140.0000 103.6139 132.1970 166.4127 195.9701  19.8750  70.0000
}

}
