% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_rna-sga-coma.R
\name{rna_experimental}
\alias{rna_experimental}
\alias{rna_rec}
\alias{rna_prior}
\title{RNA – Experimental Versions}
\usage{
rna_rec(
  total_cost,
  A,
  bounds = NULL,
  unit_costs = rep(1, length(A)),
  cmp = .Primitive(">=")
)

rna_prior(
  total_cost,
  A,
  bounds = NULL,
  check = NULL,
  cmp = .Primitive(">="),
  details = FALSE
)
}
\arguments{
\item{total_cost}{(\code{numeric(1)})\cr total survey cost \eqn{c}.
Must be strictly positive.
Additionally:
\itemize{
\item If one-sided lower bounds \eqn{m_1, \ldots, m_H} are imposed, it is
required that \eqn{c \geq \sum_{h=1}^H c_h m_h}, i.e.
\code{total_cost >= sum(unit_costs * bounds)}.
\item If one-sided upper bounds \eqn{M_1, \ldots, M_H} are imposed, it is
required that \eqn{c \leq \sum_{h=1}^H c_h M_h}, i.e.
\code{total_cost <= sum(unit_costs * bounds)}.
}}

\item{A}{(\code{numeric})\cr population constants \eqn{A_1,\ldots,A_H}.
All values must be strictly positive.}

\item{bounds}{(\code{numeric} or \code{NULL})\cr optional lower bounds \eqn{m_1,\ldots,m_H},
or upper bounds \eqn{M_1,\ldots,M_H}, or \code{NULL} to indicate that no inequality
constraints are imposed.
If not \code{NULL}, \code{bounds} is interpreted as:
\itemize{
\item lower bounds, if \code{cmp = .Primitive("<=")}, or
\item upper bounds, if \code{cmp = .Primitive(">=")}.
}

See also \code{total_cost}.}

\item{unit_costs}{(\code{numeric})\cr costs \eqn{c_1,\ldots,c_H} of surveying one
element in each stratum. Strictly positive values. May also be of length 1,
in which case the value is recycled to match the length of \code{bounds}.}

\item{cmp}{(\code{function})\cr a binary comparison operator used to check for
violations of \code{bounds}.
Must be either \code{.Primitive("<=")} (treating \code{bounds} as lower bounds and
invoking the LRNA algorithm) or \code{.Primitive(">=")} (treating \code{bounds} as
upper bounds and invoking the RNA algorithm).

The value of this argument has no effect if \code{bounds} is \code{NULL}.}

\item{check}{(\code{integer})\cr indices of strata for which allocation constraints
may be violated. For all other strata, constraint violations are not checked.}

\item{details}{(\code{logical(1)})\cr should detailed information on stratum
assignments (either take-Neyman or take-bound), values of the set function
\eqn{s}, and the number of iterations be included in the output?}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Experimental variants of the Recursive Neyman Algorithm (RNA).
}
\section{Functions}{
\itemize{
\item \code{rna_rec()}: Recursive implementation of the RNA.

\item \code{rna_prior()}: A variant of the Recursive Neyman Algorithm (RNA) that uses prior information
about strata for which allocation constraints may be violated. For all other
strata, allocations are assumed to satisfy the bounds.

}}
\note{
This code has not been extensively tested and may change in future
releases.
}
\examples{
A <- c(3000, 4000, 5000, 2000)
M <- c(100, 90, 70, 80) # upper bounds

# experimental function (not exported) – examples skipped
\dontrun{
rna_rec(total_cost = 190, A = A, bounds = M)
rna_rec(total_cost = 312, A = A, bounds = M)
rna_rec(total_cost = 339, A = A, bounds = M)
rna_rec(total_cost = 340, A = A, bounds = M)
}

}
