% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms_onesided.R
\name{rna_prior}
\alias{rna_prior}
\title{rna version that uses prior information about violations}
\usage{
rna_prior(n, a, bounds, upper = TRUE, check = NULL, assignments = FALSE)
}
\arguments{
\item{n}{(\code{number})\cr total sample size. A strictly positive scalar.}

\item{a}{(\code{numeric})\cr parameters \eqn{a_1, ..., a_H} of variance function
\eqn{D}. Strictly positive numbers.}

\item{bounds}{(\code{numeric} or \code{NULL}) \cr optional one-sided lower or upper
bounds constraints imposed on strata sample sizes. If \code{bounds} is not
\code{NULL}, it is required that \code{n <= sum(bounds)} in case of \code{upper = TRUE},
and \code{n >= sum(bounds)}, in case of \code{upper = FALSE}. Strictly positive
numbers.}

\item{upper}{(\code{flag}) \cr should values of \code{bounds} be treated as one-sided
upper bounds constraints (default)? Otherwise, they are treated as lower
bounds.}

\item{check}{(\code{integer})\cr strata indices for which the allocation can possible
be violated. For other strata allocation cannot be violated.}

\item{assignments}{(\code{flag}) \cr should information about strata assignments
(either to take-Neyman or take-bound) be added to the output?}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This is the version of the RNA that makes use of additional information
about strata for which the allocation can possibly be violated. For all other
strata allocation will not be violated.
}
