% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms_onesided.R
\name{rna_s}
\alias{rna_s}
\title{rna version that computes values of set function s}
\usage{
rna_s(n, a, bounds = NULL, upper = TRUE, assignments = FALSE)
}
\arguments{
\item{n}{(\code{number})\cr total sample size. A strictly positive scalar.}

\item{a}{(\code{numeric})\cr parameters \eqn{a_1, ..., a_H} of variance function
\eqn{D}. Strictly positive numbers.}

\item{bounds}{(\code{numeric} or \code{NULL}) \cr optional one-sided lower or upper
bounds constraints imposed on strata sample sizes. If \code{bounds} is not
\code{NULL}, it is required that \code{n <= sum(bounds)} in case of \code{upper = TRUE},
and \code{n >= sum(bounds)}, in case of \code{upper = FALSE}. Strictly positive
numbers.}

\item{upper}{(\code{flag}) \cr should values of \code{bounds} be treated as one-sided
upper bounds constraints (default)? Otherwise, they are treated as lower
bounds.}

\item{assignments}{(\code{flag}) \cr should information about strata assignments
(either to take-Neyman or take-bound) be added to the output?}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
a <- c(3000, 4000, 5000, 2000)
bounds <- c(100, 90, 70, 80)
rna_s(n = 190, a = a, bounds = bounds)

}
