% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_rnabox.R
\name{rnabox}
\alias{rnabox}
\title{Recursive Neyman Algorithm for Optimum Sample Allocation under Box
Constraints (RNABOX)}
\usage{
rnabox(
  n,
  A,
  bounds_inner = NULL,
  bounds_outer = NULL,
  cmp_inner = .Primitive(">="),
  cmp_outer = .Primitive("<=")
)
}
\arguments{
\item{n}{(\code{integerish(1)})\cr total sample size.
Must satisfy \code{n > 0}.
Additionally:
\itemize{
\item If \code{bounds_inner} is not \code{NULL}, then
\code{n >= sum(bounds_inner)} when \code{bounds_inner} are treated as lower bounds, or
\code{n <= sum(bounds_inner)} when treated as upper bounds.
\item If \code{bounds_outer} is not \code{NULL}, then
\code{n >= sum(bounds_outer)} when \code{bounds_outer} are treated as lower bounds, or
\code{n <= sum(bounds_outer)} when treated as upper bounds.
}}

\item{A}{(\code{numeric})\cr population constants \eqn{A_1,\ldots,A_H}.
All values must be strictly positive.}

\item{bounds_inner}{(\code{numeric} or \code{NULL})\cr optional bounds on sample sizes
in strata for the \strong{interim (inner) allocation} phase of \code{rnabox()}.
These can be either lower bounds \eqn{m_1,\ldots,m_H} or upper bounds
\eqn{M_1,\ldots,M_H}, depending on the value of \code{cmp_inner}.
If \code{bounds_inner} is \code{NULL}, no bounds are imposed during the interim
allocation phase.

If both \code{bounds_inner} and \code{bounds_outer} are not \code{NULL}, the following
element-wise relationships must hold:
\itemize{
\item If \code{bounds_inner} are treated as lower bounds, then \code{bounds_inner < bounds_outer}.
\item If \code{bounds_inner} are treated as upper bounds, then \code{bounds_inner > bounds_outer}.
}

\code{bounds_inner} is passed by \code{rnabox()} to \code{\link[=rna]{rna()}} as the \code{bounds} argument,
serving as an interim allocation step applied before enforcing the outer
\code{bounds_outer}.}

\item{bounds_outer}{(\code{numeric} or \code{NULL})\cr optional bounds on sample sizes
in strata, checked during the \strong{violation check (outer)} phase of the
iterative \code{rnabox()} loop.
These can be either lower bounds \eqn{m_1,\ldots,m_H} or upper bounds
\eqn{M_1,\ldots,M_H}, depending on the value of \code{cmp_outer}.
If \code{bounds_outer} is \code{NULL}, no bounds are imposed during the outer phase.

If both \code{bounds_outer} and \code{bounds_inner} are not \code{NULL}, the following
element-wise relationships must hold:
\itemize{
\item If \code{bounds_outer} are treated as lower bounds, then \code{bounds_outer < bounds_inner}.
\item If \code{bounds_outer} are treated as upper bounds, then \code{bounds_outer > bounds_inner}.
}}

\item{cmp_inner}{(\code{function})\cr
a binary comparison operator used to check for violations of \code{bounds_inner}.
Must be either \code{.Primitive("<=")} or \code{.Primitive(">=")}.
This operator determines how \code{bounds_inner} is handled:
\itemize{
\item \code{.Primitive("<=")} treats \code{bounds_inner} as lower bounds and causes
\code{rnabox()} to apply the LRNA algorithm as an inner allocation step.
\item \code{.Primitive(">=")} treats \code{bounds_inner} as upper bounds and causes
\code{rnabox()} to apply the RNA algorithm as an inner allocation step.
}

The value of this argument has no effect if \code{bounds_inner} is \code{NULL}.

\code{cmp_inner} is passed by \code{rnabox()} to \code{\link[=rna]{rna()}} as the \code{cmp} argument,
serving as an interim allocation step applied before enforcing the outer
\code{bounds_outer}.}

\item{cmp_outer}{(\code{function})\cr
a binary comparison operator used to check for violations of \code{bounds_outer}.
It must be the logical complement of \code{cmp_inner} (up to equality):
if \code{cmp_inner = .Primitive("<=")}, then \code{cmp_outer = .Primitive(">=")},
and vice versa.
It determines how \code{bounds_outer} is handled:
\itemize{
\item \code{.Primitive("<=")} treats \code{bounds_outer} as lower bounds.
\item \code{.Primitive(">=")} treats \code{bounds_outer} as upper bounds.
}

The value of this argument has no effect if \code{bounds_outer} is \code{NULL}.

\code{cmp_outer} is provided solely for computational efficiency, as its value
is fully determined by \code{cmp_inner}.}
}
\value{
A numeric vector of optimum sample allocations in strata.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Implements the Recursive Neyman Algorithm for Optimum Sample Allocation under
Box Constraints (RNABOX), as proposed in \insertCite{rnabox;textual}{stratallo}.
The algorithm solves the following optimum allocation problem, formulated in
mathematical optimization terms:

Minimize
\deqn{f(x_1,\ldots,x_H) = \sum_{h=1}^H \frac{A^2_h}{x_h}}
over \eqn{\mathbb R_+^H}, subject to
\deqn{\sum_{h=1}^H x_h = n,}
\deqn{m_h \leq x_h \leq M_h, \qquad h = 1,\ldots,H,}
where \eqn{n > 0,\, A_h > 0,\, m_h > 0,\, M_h > 0}, such that
\eqn{m_h < M_h,\, h = 1,\ldots,H}, and
\eqn{\sum_{h=1}^H m_h \leq n \leq \sum_{h=1}^H M_h}, are given numbers.
Inequality constraints are optional and may be omitted.
}
\note{
The \code{rnabox()} function is optimized for internal use and should
typically not be called directly by users. Use \code{\link[=opt]{opt()}} instead.
}
\examples{
N <- c(454, 10, 116, 2500, 2240, 260, 39, 3000, 2500, 400)
S <- c(0.9, 5000, 32, 0.1, 3, 5, 300, 13, 20, 7)
A <- N * S
m <- c(322, 3, 57, 207, 715, 121, 9, 1246, 1095, 294) # lower bounds
M <- N # upper bounds

# Regular allocation.
n <- 6000
opt_regular <- rnabox(n, A, M, m)

# Vertex allocation.
n <- 4076
opt_vertex <- rnabox(n, A, M, m)
}
\references{
\insertRef{rnabox}{stratallo}
}
\seealso{
\code{\link[=opt]{opt()}}, \code{\link[=optcost]{optcost()}}, \code{\link[=rna]{rna()}}, \code{\link[=sga]{sga()}}, \code{\link[=sgaplus]{sgaplus()}}, \code{\link[=coma]{coma()}}
}
