% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{round_oric}
\alias{round_oric}
\title{Optimal Rounding under Integer Constraints}
\usage{
round_oric(x)
}
\arguments{
\item{x}{(\code{numeric})\cr a numeric vector.}
}
\value{
An integer vector.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
x <- c(4.5, 4.1, 4.9)
round_oric(x) # 4 4 5
}
\references{
Cont, R., Heidari, M. (2014).
Optimal rounding under integer constraints.
\doi{10.48550/arXiv.1501.00014}
}
