% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding.R
\name{rounding}
\alias{rounding}
\alias{round_ran}
\alias{round_oric}
\title{Rounding of Numbers}
\usage{
round_ran(x)

round_oric(x)
}
\arguments{
\item{x}{(\code{numeric})\cr numeric vector.}
}
\value{
An integer vector.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\section{Functions}{
\itemize{
\item \code{round_ran()}: Random rounding of numbers.
A number \eqn{x} is rounded to an integer \eqn{y} according to the following
rule:

\deqn{y = \left\lfloor x \right\rfloor +
  I\!\left(u < x - \left\lfloor x \right\rfloor\right),}
where the indicator function \eqn{I : \{FALSE,\, TRUE\} \to \{0,\, 1\}} is
defined as
\deqn{
  I(b) :=
  \begin{cases}
    0, & b \text{ is } FALSE, \\
    1, & b \text{ is } TRUE,
  \end{cases}
}
and \eqn{u} is a random number drawn from the \eqn{\mathrm{Uniform}(0, 1)}
distribution.

\item \code{round_oric()}: Optimal rounding under integer constraints, as proposed by
\insertCite{Cont;textual}{stratallo}.

}}
\examples{
x <- c(4.5, 4.1, 4.9)

set.seed(5)
round_ran(x) # 5 4 4

set.seed(6)
round_ran(x) # 4 4 5

round_oric(x) # 4 4 5

}
\references{
\insertRef{Cont}{stratallo}
}
