% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratallo-package.R
\docType{package}
\name{stratallo}
\alias{stratallo-package}
\alias{stratallo}
\title{Optimum Sample Allocation in Stratified Sampling}
\description{
Optimum Sample Allocation in Stratified Sampling
}
\references{
\insertRef{stratallo_src}{stratallo}

\insertRef{rnabox_perf}{stratallo}

\insertRef{WWW}{stratallo}

\insertRef{WojciakLRNA}{stratallo}

\insertRef{WojciakMsc}{stratallo}

\insertRef{WojciakPhD}{stratallo}

\insertRef{SG}{stratallo}

\insertRef{Sarndal}{stratallo}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wwojciech/stratallo}
  \item Report bugs at \url{https://github.com/wwojciech/stratallo/issues}
}

}
\author{
Wojciech Wójciak \email{wojciech.wojciak@gmail.com}
}
\keyword{package}
