% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_other_authors.R
\name{fpia}
\alias{fpia}
\alias{fpia2}
\alias{glambda}
\alias{philambda}
\title{Fixed-Point Iteration Algorithm}
\usage{
fpia(
  n,
  Ah,
  mh = NULL,
  Mh = NULL,
  lambda0 = NULL,
  maxiter = 100,
  tol = .Machine$double.eps * 1000
)

fpia2(v0, Nh, Sh, mh = NULL, Mh = NULL, lambda0 = NULL, maxiter = 100)

glambda(lambda, n, Ah, mh = NULL, Mh = NULL)

philambda(lambda, n, Ah, mh = NULL, Mh = NULL)
}
\arguments{
\item{n}{(\code{integerish(1)})\cr total sample size.}

\item{Ah}{(\code{numeric})\cr products of population stratum sizes and standard
deviations of the study variable, \eqn{A_h = N_h S_h}.}

\item{mh}{(\code{numeric} or \code{NULL})\cr lower bounds on stratum sample sizes (optional).}

\item{Mh}{(\code{numeric} or \code{NULL})\cr upper bounds on stratum sample sizes (optional).}

\item{lambda0}{(\code{numeric(1)})\cr initial value of the parameter \eqn{\lambda} (optional).}

\item{maxiter}{(\code{integerish(1)})\cr maximum number of iterations.}

\item{tol}{(\code{numeric(1)})\cr desired convergence tolerance.}

\item{v0}{variance}

\item{Nh}{(\code{numeric})\cr population sizes in strata.}

\item{Sh}{(\code{numeric})\cr standard deviations of the study variable in strata.}

\item{lambda}{(\code{numeric(1)})\cr \eqn{\lambda}.}
}
\value{
A list with elements:
\describe{
\item{nh}{Vector of optimal allocation sizes.}
\item{iter}{Number of iterations performed.}
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Algorithm for optimum sample allocation in stratified sampling under lower-
and upper-bound constraints, based on fixed-point iteration.
}
\section{Functions}{
\itemize{
\item \code{fpia()}: 

\item \code{fpia2()}: Variant of \code{fpia()} using variance-based parametrization.

\item \code{glambda()}: Helper function for the \code{fpia()}

\item \code{philambda()}: Helper function for the \code{fpia()}.

}}
\references{
\insertRef{MSW}{stratallo}
}
