% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.R
\docType{data}
\name{pop10_mM}
\alias{pop10_mM}
\title{Example Population with 10 Strata and Lower and Upper Bounds}
\format{
A matrix with 10 rows and 5 columns:
\code{N} (strata sizes), \code{S} (standard deviations of a
variable under study in strata), \code{m} (lower bounds), \code{M} (upper bounds).
}
\usage{
pop10_mM
}
\description{
Artificial example population with 10 strata defined by strata
sizes and standard deviations of some study variable. Additionally, the
lower and upper bounds have been specified for the sample in strata.
}
\keyword{datasets}
