% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.R
\docType{data}
\name{pop507}
\alias{pop507}
\title{Example Population with 507 Strata}
\format{
A matrix with 507 rows and 2 columns: \code{N} with strata sizes and \code{S}
standard deviations of a variable under study in strata.
}
\usage{
pop507
}
\description{
Artificial example population with 507 strata defined by strata
sizes and standard deviations of some study variable.
}
\keyword{datasets}
