% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop507}
\alias{pop507}
\title{Example Population with 507 Strata}
\format{
A matrix with 507 rows and 3 variables:
\describe{
\item{N}{stratum size}
\item{S}{standard deviation of study variable in stratum}
\item{unit_cost}{cost of surveying one element in stratum}
}
}
\usage{
pop507
}
\description{
A dataset containing the artificial population with 507 strata.
}
\keyword{datasets}
