% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.R
\docType{data}
\name{pop969}
\alias{pop969}
\title{Example Population with 969 Strata}
\format{
A matrix with 969 rows and 2 columns: \code{N} with strata sizes and \code{S}
standard deviations of a variable under study in strata.
}
\usage{
pop969
}
\description{
Artificial example population with 969 strata defined by strata
sizes and standard deviations of some study variable.
}
\keyword{datasets}
